dialogue_digital/requirements.txt
websocket-client==1.8.0
numpy
pypinyin
pyaudio